// frontend/src/clients/core/types.gen.ts
export type DataType = 'string' | 'string_array' | 'int' | 'float' | 'boolean' | 'date';
// ...
export type GlobalMetadataOutput = {
    /**
     * Id
     */
    id?: string | null;
    source: MetadataSource;
    /**
     * Name
     */
    name: string;
    data_type: DataType;
    /**
     * Formatter
     */
    formatter?: string | null;
    /**
     * Links To Document Id
     */
    links_to_document_id?: boolean;
    /**
     * Links To Metadata Id
     */
    links_to_metadata_id?: string | null;
    linked_to?: GlobalMetadataOutput | unknown;
};

/**
 * GlobalMetadataAvailable
 */
export type GlobalMetadataAvailable = {
    /**
     * Source
     */
    source: string;
    /**
     * Key
     */
    key: string;
};

/**
 * GlobalMetadataAvailableList
 */
export type GlobalMetadataAvailableList = Array<GlobalMetadataAvailable>;

/**
 * GlobalMetadataBySource
 */
export type GlobalMetadataBySource = {
    /**
     * Documents
     */
    documents: {
        [key: string]: GlobalMetadataLinkedSource;
    };
    /**
     * Annotations
     */
    annotations: {
        [key: string]: GlobalMetadataLinkedSource;
    };
    /**
     * Sentences
     */
    sentences: {
        [key: string]: GlobalMetadataLinkedSource;
    };
};

/**
 * GlobalMetadataLinkedSource
 */
export type GlobalMetadataLinkedSource = {
    /**
     * Id
     */
    id: string;
    source: MetadataSource;
    /**
     * Name
     */
    name: string;
    data_type: DataType;
    /**
     * Formatter
     */
    formatter?: string | null;
    /**
     * Links To Document Id
     */
    links_to_document_id?: boolean;
    /**
     * Links To Metadata Id
     */
    links_to_metadata_id?: string | null;
    linked_to?: GlobalMetadataOutput | unknown;
    /**
     * Linked Name
     */
    linked_name?: string | null;
    linked_source?: MetadataSource | unknown;
};
// ...
export type MetadataSource = 'document' | 'annotation' | 'sentence';